requirements.txt
tensorflow==2.15.0
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
numpy==1.26.4
Pillow==10.3.0
scikit-learn==1.4.2
matplotlib==3.8.4
seaborn==0.13.2
pandas==2.2.2
streamlit==1.35.0
locust==2.24.1
